Chatlist, ChatRoom OK

+++ backend/src/controllers/memberAuth.js
@@ -9,20 +9,20 @@ require("dotenv").config(); // 加了這行就可以抓到 mailer
 
 // 使用者輸入信箱後，判斷是否是會員，回傳一個狀態（若是進入登入頁面，否則進入註冊）
 const LOR = async (req, res) => {
-  const email = req.body.email;
+  const { email } = req.query;
   // console.log(email);
   if (!email) {
     return res.status(400).json({ error: "Email cannot be empty" });
   }
   try {
     const existingUser = await MemberAuth.findOne({ email });
-    // console.log(existingUser);
-    if (existingUser) {
+    console.log(existingUser.password);
+    if (existingUser && existingUser.password) {
       // If user exists
-      return res.json({ status: "success" });
+      return res.status(200).json({ status: "success" });
     } else {
       // If user doesn't exist
-      return res.json({ status: "None" });
+      return res.status(404).json({ status: "None" });
     }
   } catch (error) {
     return res.status(500).json({ error: "Internal server error" });
@@ -53,20 +53,26 @@ const login = async (req, res) => {
     });
     // Check if user exists
     if (!user) {
-      return res.json({
+      return res.status(404).json({
         status: "error",
         message: "Email不存在",
       });
+    } else if (!user.password) {
+      return res.status(404).json({
+        status: "error",
+        message: "請進行驗證碼驗證",
+      });
     }
+    console.log(user);
     // Check if password correct
     const passwordMatch = await bcrypt.compare(data.password, user.password);
     if (passwordMatch) {
       const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET_KEY, {
         expiresIn: "1h",
       });
-      res.cookie("token", token, { httpOnly: true });
+      res.cookie("token", token, { httpOnly: true, secure: true });
 
-      return res.json({
+      return res.status(200).json({
         status: "success",
         message: "登入成功",
         data: {
@@ -76,7 +82,7 @@ const login = async (req, res) => {
         },
       });
     } else {
-      return res.json({
+      return res.status(400).json({
         status: "error",
         message: "密碼錯誤",
       });
@@ -112,7 +118,6 @@ const registerMember = async (req, res) => {
   // create member and memberAuth in database
   if (!user) {
     const member = await Member.create({ username: email.split("@")[0] });
-    console.log(member);
     await MemberAuth.create({
       user_id: member._id,
       email: email,
@@ -122,7 +127,7 @@ const registerMember = async (req, res) => {
   } else if (user && !user.password) {
     await MemberAuth.updateOne({ email }, { verification_code: code });
   } else {
-    return res.json({
+    return res.status(400).json({
       status: "error",
       message: "已成為會員，請登入",
     });
@@ -137,7 +142,7 @@ const registerMember = async (req, res) => {
       text: `Your verification code is: ${code}`,
     });
 
-    res.json({
+    res.status(200).json({
       status: "verified",
       message: "驗證碼已寄出",
     });
@@ -177,7 +182,7 @@ const verifyRegisterMember = async (req, res) => {
       { email },
       { is_verified: true }
     );
-    res.json({
+    res.status(200).json({
       status: "verified",
       message: "電子郵件驗證成功",
       // data: updatedUser,
@@ -188,21 +193,47 @@ const verifyRegisterMember = async (req, res) => {
   }
 };
 
+//檢查使用者名稱有無重複
+const checkUsername = async (req, res) => {
+  const { username, email } = req.query;
+  const m0 = await Member.findOne({ username: username });
+  const m1 = await MemberAuth.findOne({ email: email });
+  if (!m0) {
+    return res.status(200).json({
+      status: "success",
+      message: "可使用此名稱",
+    });
+  } else {
+    if (m0._id.toString() == m1.user_id.toString()) {
+      return res.status(200).json({
+        status: "success",
+        message: "可使用此名稱",
+      });
+    }
+  }
+  return res.status(409).json({
+    status: "failed",
+    message: "使用者名稱已有人使用，請更換其他名稱",
+  });
+};
+
 //設定密碼、使用者名稱、交換學校名稱
 const verifiedMember = async (req, res) => {
   const { email, password, username, exchange_school_name } = req.body;
 
   try {
+    //設定密碼
     const hashedPassword = await hashPassword(password);
     const user = await MemberAuth.findOne({ email });
-
+    console.log("user:", user);
     if (user.password) {
-      return res.json({
+      return res.status(400).json({
         status: "failed",
         message: "使用者已設定過密碼，請登入",
         data: email,
       });
     } else if (user) {
+      //未設定過密碼，設定member auth和member
       const ma = await MemberAuth.findOneAndUpdate(
         { email },
         { password: hashedPassword },
@@ -216,7 +247,7 @@ const verifiedMember = async (req, res) => {
         },
         { new: true }
       );
-      return res.json({
+      return res.status(200).json({
         status: "success",
         message: "註冊成功！",
         data: m,
@@ -224,7 +255,7 @@ const verifiedMember = async (req, res) => {
     }
 
     // If user doesn't exist, return failed message
-    return res.json({
+    return res.status(404).json({
       status: "failed",
       message: "使用者不存在！",
       data: email,
@@ -239,27 +270,25 @@ const deleteTestMember = async (req, res) => {
   const { email } = req.body; // Extract email from request body
   try {
     // Assuming MemberAuth is your Mongoose model
-    const toDelete = await MemberAuth.findOne({ email });
-    const deletedMemberAuth = await MemberAuth.deleteOne({ email });
-    const deletedMember = await Member.deleteOne({
-      //default
-      username: email.split("@")[0],
-    });
+    const m = await MemberAuth.findOne({ email });
+    if (m) {
+      let memberId = m.user_id;
+      const deletedMemberAuth = await MemberAuth.deleteOne({ email });
+      const deletedMember = await Member.deleteOne({
+        _id: memberId,
+      });
+      console.log(`${deletedMember.deletedCount} member(s) deleted.`);
+      console.log(`${deletedMemberAuth.deletedCount} member(s) auth deleted.`);
 
-    if (
-      deletedMemberAuth.deletedCount === 0 ||
-      deletedMember.deletedCount === 0
-    ) {
-      // If no documents were deleted
-      return res.status(404).json({
-        status: "error",
-        message: "No matching member found for deletion",
+      return res.status(200).json({
+        status: "success",
+        message: "Deleted successfully",
+        data: email,
       });
     }
-
-    return res.json({
-      status: "success",
-      message: "Deleted successfully",
+    return res.status(404).json({
+      status: "failed",
+      message: "no existing user",
       data: email,
     });
   } catch (error) {
@@ -273,6 +302,7 @@ module.exports = {
   login,
   registerMember,
   verifyRegisterMember,
+  checkUsername,
   verifiedMember,
   deleteTestMember,
 };
